+++ esmerald/routing/apis/base.py
@@ -9,7 +9,7 @@ from typing_extensions import Annotated, Doc
 from esmerald.permissions.utils import is_esmerald_permission, wrap_permission
 
 if TYPE_CHECKING:  # pragma: no cover
-    from esmerald.interceptors.types import Interceptor
+    from esmerald.core.interceptors.types import Interceptor
     from esmerald.openapi.schemas.v3_1_0.security_scheme import SecurityScheme
     from esmerald.permissions.types import Permission
     from esmerald.routing.gateways import Gateway, WebSocketGateway